requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
supabase==1.0.3
python-dotenv==1.0.0